fix: 🐛 Removed react-native-fs, audit fix
BREAKING CHANGE: 🧨 android: This release introduces the setSupportMultipleWindows prop for
Android. This sets the underlying Android WebView setting
setSupportMultipleWindows. This prop defaults to true (previously
false), and serves to mitigate the security advisory. The primary way
this new behavior changes existing React Native WebView implementations
on Android is that links that open in new tabs/windows (such as <a
target="_blank">) will now prompt to open in the system browser, rather
than re-using the current WebView.